requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
async-timeout==5.0.1
attrs==24.3.0
bidict==0.23.1
certifi==2024.8.30
charset-normalizer==3.4.0
frozenlist==1.5.0
h11==0.14.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.32.3
simple-websocket==1.1.0
typing_extensions==4.12.2
urllib3==2.2.3
wsproto==1.2.0
yarl==1.18.3
